+++ src/models/bracket_races.rs
@@ -2,27 +2,21 @@ use crate::models::bracket_race_infos::BracketRaceInfo;
 use crate::models::bracket_rounds::BracketRound;
 use crate::models::brackets::Bracket;
 use crate::models::player::Player;
-use crate::models::season::Season;
 use crate::schema::bracket_races;
 use crate::update_fn;
 use crate::utils::format_hms;
+use crate::BracketRaceState;
 use crate::BracketRaceStateError;
 use crate::NMGLeagueBotError;
 use chrono::{DateTime, Duration, TimeZone};
 use diesel::prelude::*;
 use diesel::SqliteConnection;
-use log::warn;
 use serde::Serialize;
 use std::collections::HashMap;
 use std::fmt::{Display, Formatter};
 use swiss_pairings::MatchResult;
 
-#[derive(serde::Serialize, serde::Deserialize, Eq, PartialEq, Debug)]
-pub enum BracketRaceState {
-    New,
-    Scheduled,
-    Finished,
-}
+
 
 #[derive(serde::Serialize, serde::Deserialize)]
 pub enum PlayerResult {
@@ -103,6 +97,21 @@ impl BracketRace {
             .load(conn)
             .map_err(From::from)
     }
+
+    pub fn get_unfinished_races_for_player(
+        player: &Player,
+        conn: &mut SqliteConnection,
+    ) -> Result<Vec<BracketRace>, NMGLeagueBotError> {
+        bracket_races::table
+            .filter(bracket_races::state.ne(serde_json::to_string(&BracketRaceState::Finished)?))
+            .filter(
+                bracket_races::player_1_id
+                    .eq(player.id)
+                    .or(bracket_races::player_2_id.eq(player.id)),
+            )
+            .load(conn)
+            .map_err(From::from)
+    }
 }
 
 impl BracketRace {
@@ -149,6 +158,10 @@ impl BracketRace {
         Ok((p1, p2))
     }
 
+    pub fn involves_player(&self, player: &Player) -> bool {
+        self.player_1_id == player.id || self.player_2_id == player.id
+    }
+
     pub fn bracket(&self, conn: &mut SqliteConnection) -> Result<Bracket, diesel::result::Error> {
         Bracket::get_by_id(self.bracket_id, conn)
     }
@@ -170,7 +183,10 @@ impl BracketRace {
         match self.state()? {
             BracketRaceState::New | BracketRaceState::Scheduled => {}
             BracketRaceState::Finished => {
-                return Err(BracketRaceStateError::InvalidState);
+                return Err(BracketRaceStateError::InvalidState(
+                    vec![BracketRaceState::New, BracketRaceState::Scheduled],
+                    BracketRaceState::Finished,
+                ));
             }
         };
         let mut info = self.info(conn)?;
@@ -193,8 +209,12 @@ impl BracketRace {
         p2: Option<&PlayerResult>,
         force: bool,
     ) -> Result<(), BracketRaceStateError> {
-        if !force && self.state()? == BracketRaceState::Finished {
-            return Err(BracketRaceStateError::InvalidState);
+        let state = self.state()?;
+        if !force && state == BracketRaceState::Finished {
+            return Err(BracketRaceStateError::InvalidState(
+                vec![BracketRaceState::Finished],
+                state,
+            ));
         }
         if let Some(p1r) = p1 {
             self.player_1_result = Some(serde_json::to_string(p1r)?);
@@ -224,7 +244,7 @@ impl BracketRace {
         let (p1, p2) = match (self.player_1_result(), self.player_2_result()) {
             (Some(p1r), Some(p2r)) => (p1r, p2r),
             _ => {
-                return Err(BracketRaceStateError::InvalidState);
+                return Err(BracketRaceStateError::MissingResult);
             }
         };
 
@@ -279,46 +299,7 @@ impl BracketRace {
     update_fn! {}
 }
 
-pub fn get_current_round_race_for_player(
-    player: &Player,
-    conn: &mut SqliteConnection,
-) -> Result<Option<BracketRace>, diesel::result::Error> {
-    let sn = match Season::get_active_season(conn)? {
-        Some(s) => s,
-        None => {
-            return Ok(None);
-        }
-    };
-
-    for bracket in sn.brackets(conn)? {
-        let round = match bracket.current_round(conn)? {
-            Some(r) => r,
-            None => {
-                continue;
-            }
-        };
-        let mut races: Vec<BracketRace> = bracket_races::table
-            .filter(bracket_races::round_id.eq(round.id))
-            .filter(
-                bracket_races::player_1_id
-                    .eq(player.id)
-                    .or(bracket_races::player_2_id.eq(player.id)),
-            )
-            .load(conn)?;
-        if races.is_empty() {
-            continue;
-        } else {
-            if races.len() != 1 {
-                warn!("Multiple races for same racer?");
-            }
-            return Ok(races.pop());
-        }
-    }
-
-    Ok(None)
-}
-
-#[derive(Insertable)]
+#[derive(Insertable, Debug)]
 #[diesel(table_name=bracket_races)]
 pub struct NewBracketRace {
     bracket_id: i32,